requirements.txt
django 
djangorestframework 
google-adk 
python-dotenv 
daphne 
django-jazzmin
